+++ components/modules/survey-home/components/survey-home-detail/index.ts
@@ -1,3 +1,1 @@
-export * from "./survey-home-detail.url-button";
-export * from "./survey-home-detail.start-button";
-export * from "./survey-home-detail.statistics"
+export * from "./survey-home-detail.components";
