--- components/modules/survey-home/components/survey-home-detail/survey-home-detail.statistics.tsx
@@ -1,13 +1,0 @@
-import {FC} from 'react';
-import { Typography } from 'antd';
-
-type ISurveyHomeDetailStatisticsProp = {};
-
-export const SurveyHomeDetailStatistics: FC<ISurveyHomeDetailStatisticsProp> = () => {
-    const { Title } = Typography;
-    return (
-        <>
-        <Title level={4} style={{textAlign:"center"}} mark>총 7명이 방문했어요!</Title>
-        </>
-    );
-}
