+++ screens/index.ts
@@ -1,2 +1,1 @@
 export * from './survey';
-export * from './survey-home';
