--- screens/survey-home/index.ts
@@ -1,1 +1,0 @@
-export * from './survey-home-processing/survey-home-processing.screen'
