--- screens/survey-home/survey-home-processing/index.ts
@@ -1,1 +1,0 @@
-export * from './survey-home-processing.screen'
