+++ screens/survey/index.ts
@@ -1,1 +1,2 @@
 export * from "./processing/survey-processing.screen";
+export * from "./home/survey-home.screen";
